Rearrange module structure and split up utils.py - make utils a folder - create the constants module - create the widgets module - updated imports
