+++ spydertop/screens/help.py
@@ -17,7 +17,8 @@ from asciimatics.event import KeyboardEvent
 
 from spydertop.model import AppModel
 from spydertop.widgets import FuncLabel, Padding
-from spydertop.utils import ExtendedParser, add_palette
+from spydertop.utils import add_palette
+from spydertop.utils.types import ExtendedParser
 
 if sys.version_info >= (3, 8):
     from importlib import metadata
