+++ spydertop/utils/cursorlist.py
